Moved some helper functions to a shared helper.
